# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/webhooks": {
      "post": {
        "operationId": "create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWebhookEndpointDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedWebhookEndpointResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Register an endpoint; the signing secret is returned once",
        "tags": [
          "Webhooks"
        ]
      },
      "get": {
        "operationId": "list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WebhookEndpointResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "List endpoints",
        "tags": [
          "Webhooks"
        ]
      }
    },
    "/v1/webhooks/{id}": {
      "delete": {
        "operationId": "remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Remove an endpoint and its delivery history",
        "tags": [
          "Webhooks"
        ]
      }
    },
...
  },
  "info": {
    "title": "Dunlin",
    "description": "Multi-tenant invoicing, dunning and payment reconciliation. Authenticate with a tenant API key as a Bearer token; tenant creation uses the platform token.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "api-key": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Tenant API key"
      },
      "platform": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Platform-Token"
      }
    },
    "schemas": {
...
      "CreateWebhookEndpointDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "example": "https://erp.example.com/hooks/dunlin"
          },
          "events": {
            "type": "array",
            "example": [
              "invoice.paid",
              "dunning.notice_created"
            ],
            "items": {
              "type": "string",
              "enum": [
                "invoice.issued",
                "invoice.paid",
                "invoice.voided",
                "payment.recorded",
                "dunning.notice_created",
                "bank_import.completed",
                "*"
              ]
            }
          }
        },
        "required": [
          "url",
          "events"
        ]
      },
      "CreatedWebhookEndpointResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "events": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "active": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string"
          },
          "secret": {
            "type": "string",
            "description": "HMAC secret for X-Dunlin-Signature. Shown once."
          }
        },
        "required": [
          "id",
          "url",
          "events",
          "active",
          "createdAt",
          "secret"
        ]
      },
      "WebhookEndpointResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "events": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "active": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "url",
          "events",
          "active",
          "createdAt"
        ]
      },
...
    }
  }
}